Create TACIT_HC-SR04
This is the code for Tacit glove made by steve hoefer(@grathio). But instead of using the costly PING sensors we have used the cheaper HC-SR04 sensors. and the required changes are made to make it work with the same. this code is built from the scratch and everyone is welcomed to make changes and improve it further. I haven't included the flicker reduction algorithm due to some issues but i will improve it further in some time.